separate logic into separate file
